# server/api.yaml
...
openapi: 3.0.3
info:
  title: Dojo API
  description: |-
    Welcome to the Dojo API specification
  version: 1.0.0
...
servers:
  - url: http://localhost:7777/api/
    description: Local development server
  - url: https://dojo-test.hackyourfuture.net/api/
    description: Test server
...
paths:
...
  /search:
    get:
      tags:
        - search
      summary: Find trainees
      description:  Find trainees by a keyword. The search will try to match the first name, last name or email.
      security:
        - cookie_token:
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Limit the number of results. Default value is 20, Max value is 50.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  hits:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: 'Pv3R416r'
                            name:
                              type: string
                              example: John Doe
                      size:
                        type: integer
                        example: 45

        '400':
          description: Missing query
  /trainees:
    post:
      tags:
        - trainees
      summary: Create new profile
      description: ''
      security:
        - cookie_token:
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trainee'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trainee'
...
  /dashboard:
      get:
        tags:
          - dashboard
        summary: Get the data to display on the dashboard
        description: ''
        security:
          - cookie_token:
        parameters:
        - name: startDate
          in: query
          description: Include all trainees who started the program on or after startDate
          required: true
          schema:
            type: string
            format: date-time
            example: '2024-01-01'

        - name: endDate
          in: query
          description: Include all trainees who started the program before or on endDate
          required: true
          schema:
            type: string
            format: date-time
            example: '2025-01-01'

        responses:
          200:
            description: successful operation
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    demographics:
                      type: object
                      properties:
                        genderDistribution:
                          type: array
                          items: 
                            $ref: '#/components/schemas/DashboardData'
                        countryOfOrigin:
                          type: array
                          items: 
                            $ref: '#/components/schemas/DashboardData'
                    program:
                      properties:
                        graduations:
                          type: array
                          items: 
                            $ref: '#/components/schemas/DashboardData'
                        employment:
                          type: array
                          items: 
                            $ref: '#/components/schemas/DashboardData'
components:
  schemas:
...
    Trainee:
      type: object
      required:
        - personalInfo
        - contactInfo
        - educationInfo
      properties:
        id:
          type: string
          example: '65ea2e1e696deb3b3ebff843'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
        displayName:
          type: string
          example: Johnny Doe
        personalInfo:
          $ref: '#/components/schemas/TraineePersonalInfo'
        contactInfo:
          $ref: '#/components/schemas/TraineeContactInfo'
        educationInfo:
          $ref: '#/components/schemas/TraineeEducationInfo'  
        employmentInfo:
          $ref: '#/components/schemas/TraineeEmploymentInfo'
        interactions:
          type: array
          items:
            $ref: '#/components/schemas/TraineeInteraction'
    TraineePersonalInfo:
      type: object
      required:
        - firstName
        - lastName
        - gender
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        preferredName:
          type: string
          example: Johnny
        gender:
          type: string
          enum:
            - man
            - woman
            - non-binary
            - other
          example: male
        pronouns:
          type: string
          example: He/His
        location:
          type: string
          example: Amsterdam
        englishLevel:
          type: string
          enum:
            - needs-work
            - moderate
            - good
          example: good
        professionalDutch:
            type: boolean
            example: true
        countryOfOrigin:
          type: string
          example: Syria
        background:
          type: string
          enum:
            - refugee
            - family-reunification
            - partner-of-skilled-migrant
            - vulnerable-group
            - eu-citizen
          example: refugee
        hasWorkPermit:
          type: boolean
          example: true
        residencyStatus:
          type: string
          enum:
            - first-interview
            - second-interview
            - residency
            - citizenship
        receivesSocialBenefits:
          type: boolean
          example: true
        caseManagerUrging:
          type: boolean
          example: false
        educationLevel:
          type: string
          enum:
            - none
            - high-school
            - diploma
            - bachelors-degree
            - masters-degree
          example: bachelors-degree
        educationBackground:
          type: string
          example: BSc in Biology
        comments:
          type: string
    TraineeContactInfo:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          example: john.doe@example.com
        slackId:
          type: string
          example: U068AQ9G99F
        phone:
          type: string
          example: 0612345678
        githubHandle:
          type: string
          example: johngithub
        linkedin:
          type: string
          example: https://linkedin.com/in/john.doe
    TraineeEducationInfo:
      type: object
      required:
        - learningStatus
        - startCohort
      properties:
        startCohort:
          type: number
          example: 46
          minimum: 0
          maximum: 999
        currentCohort:
          type: number
          example: 47
          minimum: 0
          maximum: 999
        learningStatus:
          type: string
          enum:
            - studying
            - graduated
            - on-hold
            - quit
          example: studying
        startDate:
          type: string
          format: date-time
          example: 2024-01-01
        graduationDate:
          type: string
          format: date-time
          example: 2024-01-01  
        quitReason:
          type: string
          enum:
            - technical
            - social-skills
            - personal
            - withdrawn
            - municipality-or-monetary
            - left-nl
            - other
        quitDate:
          type: string
          format: date-time
          example: 2024-01-01
        strikes:
          type: array
          items:
            $ref: '#/components/schemas/Strike'
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/Assignment'
        tests:
          type: array
          items:
            $ref: '#/components/schemas/Test'
        comments:
          type: string
    TraineeEmploymentInfo:
      type: object
      required:
        - jobPath
      properties:
        jobPath:
          type: string
          enum:
            - not-graduated
            - searching
            - internship
            - tech-job
            - non-tech-job
            - not-searching
            - other-studies
            - no-longer-helping
        cvURL:
          type: string
          example: https://example.com/cv.pdf
        availability:
          type: string
        preferredRole:
          type: string        
        preferredLocation:
          type: string   
        drivingLicense:
          type: boolean
          example: false
        extraTechnologies:
          type: string  
        employmentHistory:
          type: array
          items:
            $ref: '#/components/schemas/TraineeEmploymentHistory'
        comments: 
          type: string
    Strike:
      type: object
      required:
        - date
        - reason
        - reporterID
        - comments
      properties:
        id:
          type: string
          example: g5HQGuL8
        date:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00.000Z   
        reporterID:
          type: string
          example: g4G3dc4d
        reason:
          type: string
          enum:
            - absence
            - incomplete-submission
            - late-attendance
            - late-submission
            - missed-submission
            - pending-feedback
            - other
        comments: 
          type: string
    Assignment:
      type: object
      required:
        - type
        - status
      properties:
        id:
          type: string
          example: g5HQGuU8
        createDate:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00.000Z  
        type:
          type: string
          example: 'JS week 2'
        status:
          type: string
          example: 'waiting-review'
        content:
          type: string
          example: 'https://github.com/link/to/assignment'
        comments: 
          type: string
    Test:
      type: object
      required:
        - testDate
        - type
        - result
      properties:
        id:
          type: string
          example: g5HQGuL8
        date:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00.000Z  
        type:
          type: string
          enum:
            - presentation
            - javaScript
            - browsers-interview
            - using-apis-interview
            - nodejs
            - react-interview
            - final-project-interview
        grade:
          type: number
          example: 8.6
        result:
          type: string
          enum:
            - passed
            - passed-with-warning
            - failed
            - disqualified
        comments: 
          type: string
    TraineeEmploymentHistory:
      type: object
      required:
        - startDate
        - type
        - role
        - feeCollected
      properties:
        id:
          type: string
          example: g5HQGuL8
        type:
          type: string
          enum:
          - internship
          - job
        companyName:
          type: string
          example: HackYourFuture
        role:
          type: string
        startDate:
          type: string
          format: date-time
          example: 2024-01-01
        endDate:
          type: string
          format: date-time
          example: 2024-01-01 
        feeCollected:
          type: boolean
        feeAmount:
          type: number
          example: 500
        comments: 
          type: string
    TraineeInteraction:
      type: object
      required:
        - type
        - details
      properties:
        id:
          type: string
          example: g5HQGuL8
        date:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00.000Z 
        type:
          type: string
          enum:
            - call
            - feedback
            - tech-hour
            - in-person
            - other
        reporterID:
          type: string
          example: 65ea2e1e696deb3b3ebff843
        details:
          type: string
...
    DashboardData:
      type: object
      properties:
        label:
          type: string
        value: 
          type: number
          example: 42.12
        percent: 
          type: number
          example: 58.1
...
  securitySchemes:
    cookie_token:
      type: apiKey
      name: dojo_token
      description: A JWT which was provided after a successful login
      in: cookie